backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pymupdf==1.24.10
anthropic==0.34.2
pydantic==2.9.2
python-dotenv==1.0.1
